# app/backend/contracts/openapi.v0.1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "서버가 AUTH_PRINCIPALS_FILE의 SHA-256 digest로 검증하는 Bearer token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Answervice Control Plane",
    "version": "OPENAPI-v1.1.0-DRAFT"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/mcp": {
      "get": {
        "operationId": "mcpGet",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Mcp Get"
      },
      "post": {
        "operationId": "mcpPost",
        "parameters": [
          {
            "in": "header",
            "name": "MCP-Protocol-Version",
            "required": true,
            "schema": {
              "title": "Mcp-Protocol-Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Mcp-Method",
            "required": true,
            "schema": {
              "title": "Mcp-Method",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Mcp-Name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mcp-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Mcp Post"
      }
    },
...
  }
}